Add Hadran text to the end of each masechet.
Precompute the hadran API request and do some post-processing, storing it as a text file. Storing it in verison control helps to check for changes, especially with the text replacements for masechet name. Because the text/comments should rarely change, this seems fine.

In service of this, I also added vocalized hebrew names to the metadata of each of the masechtot.