Merge links requests similar to  9e8858f4f4567b326b50cd39209c2807541de739
The wins here aren't as big since links requests seem to be must cheaper, but they do still shave off somewhat significant time. Requests used in the test data saw 11-60% reduction in links requests, and requests for Genesis 1 saw about a 1s reduction in total time used for links requests (33%), which is relevant since it's a page with loads of links.