Bring back offline mode, but this time don't cache API requests through the service worker API and instead use the custom built storage cache.
This will restore offline mode but help maintain the amount of data that is cached client side.

I seem to remember that there was some weirdness that was observed with the service worker, but I don't remember what it was. Worst comes to worse this commit can also be reverted if necessary.

This partially reverts commit 058696886072700ad4a887a98b0eacfbe3396849.

I reme