Support different text versions (part 1).
This commit adds all of the support, but doesn't actually request any non-default versions. This is to make it easy to visualize that the structure of the code does not intrinsically change the expected outputs.

The very next commit will request the versions, and then we can isolate the changes just for Tanakh and Mishna.